+++ pkg/validation/options_test.go
@@ -28,7 +28,7 @@ func testOptions() *options.Options {
 		URI:  "http://127.0.0.1:8080/",
 	})
 	o.Cookie.Secret = cookieSecret
-	o.Providers[0].ProviderID = providerID
+	o.Providers[0].ID = providerID
 	o.Providers[0].ClientID = clientID
 	o.Providers[0].ClientSecret = clientSecret
 	o.EmailDomains = []string{"*"}
@@ -59,7 +59,7 @@ func TestNewOptions(t *testing.T) {
 func TestClientSecretFileOptionFails(t *testing.T) {
 	o := options.NewOptions()
 	o.Cookie.Secret = cookieSecret
-	o.Providers[0].ProviderID = providerID
+	o.Providers[0].ID = providerID
 	o.Providers[0].ClientID = clientID
 	o.Providers[0].ClientSecretFile = clientSecret
 	o.EmailDomains = []string{"*"}
@@ -97,7 +97,7 @@ func TestClientSecretFileOption(t *testing.T) {
 
 	o := options.NewOptions()
 	o.Cookie.Secret = cookieSecret
-	o.Providers[0].ProviderID = providerID
+	o.Providers[0].ID = providerID
 	o.Providers[0].ClientID = clientID
 	o.Providers[0].ClientSecretFile = clientSecretFileName
 	o.EmailDomains = []string{"*"}
@@ -115,8 +115,7 @@ func TestClientSecretFileOption(t *testing.T) {
 
 func TestGoogleAdminEmailOptions(t *testing.T) {
 	o := testOptions()
-	o.Providers[0].GoogleConfig.GoogleGroups = []string{"test_group"}
-	o.Providers[0].GoogleConfig.GoogleAdminEmail = "test@gmail.com"
+	o.Providers[0].GoogleConfig.Groups = []string{"googlegroup"}
 	err := Validate(o)
 	assert.NotEqual(t, nil, err)
 
@@ -127,8 +126,7 @@ func TestGoogleAdminEmailOptions(t *testing.T) {
 
 func TestGoogleServiceAccountJSONOptions(t *testing.T) {
 	o := testOptions()
-	o.Providers[0].GoogleConfig.GoogleGroups = []string{"test_group"}
-	o.Providers[0].GoogleConfig.GoogleServiceAccountJSON = "file_doesnt_exist.json"
+	o.Providers[0].GoogleConfig.ServiceAccountJSON = "file_doesnt_exist.json"
 	err := Validate(o)
 	assert.NotEqual(t, nil, err)
 
@@ -140,8 +138,9 @@ func TestGoogleServiceAccountJSONOptions(t *testing.T) {
 
 func TestGoogleGroupInvalidFile(t *testing.T) {
 	o := testOptions()
-	o.Providers[0].GoogleConfig.GoogleAdminEmail = "admin@example.com"
-	o.Providers[0].GoogleConfig.GoogleServiceAccountJSON = "file_doesnt_exist.json"
+	o.Providers[0].GoogleConfig.Groups = []string{"test_group"}
+	o.Providers[0].GoogleConfig.AdminEmail = "admin@example.com"
+	o.Providers[0].GoogleConfig.ServiceAccountJSON = "file_doesnt_exist.json"
 	err := Validate(o)
 	assert.NotEqual(t, nil, err)
 
@@ -242,9 +241,9 @@ func TestValidateSignatureKeyUnsupportedAlgorithm(t *testing.T) {
 
 func TestSkipOIDCDiscovery(t *testing.T) {
 	o := testOptions()
-	o.Providers[0].ProviderType = "oidc"
-	o.Providers[0].OIDCConfig.OIDCIssuerURL = "https://login.microsoftonline.com/fabrikamb2c.onmicrosoft.com/v2.0/"
-	o.Providers[0].OIDCConfig.SkipOIDCDiscovery = true
+	o.Providers[0].Type = "oidc"
+	o.Providers[0].OIDCConfig.IssuerURL = "https://login.microsoftonline.com/fabrikamb2c.onmicrosoft.com/v2.0/"
+	o.Providers[0].OIDCConfig.SkipDiscovery = true
 
 	err := Validate(o)
 	assert.Equal(t, "invalid configuration:\n"+
@@ -252,7 +251,7 @@ func TestSkipOIDCDiscovery(t *testing.T) {
 
 	o.Providers[0].LoginURL = "https://login.microsoftonline.com/fabrikamb2c.onmicrosoft.com/oauth2/v2.0/authorize?p=b2c_1_sign_in"
 	o.Providers[0].RedeemURL = "https://login.microsoftonline.com/fabrikamb2c.onmicrosoft.com/oauth2/v2.0/token?p=b2c_1_sign_in"
-	o.Providers[0].OIDCConfig.OIDCJwksURL = "https://login.microsoftonline.com/fabrikamb2c.onmicrosoft.com/discovery/v2.0/keys"
+	o.Providers[0].OIDCConfig.JwksURL = "https://login.microsoftonline.com/fabrikamb2c.onmicrosoft.com/discovery/v2.0/keys"
 
 	assert.Equal(t, nil, Validate(o))
 }
@@ -309,7 +308,7 @@ func TestProviderCAFilesError(t *testing.T) {
 	assert.NoError(t, os.Remove(file.Name()))
 
 	o := testOptions()
-	o.Providers[0].ProviderCAFiles = append(o.Providers[0].ProviderCAFiles, file.Name())
+	o.Providers[0].CAFiles = append(o.Providers[0].CAFiles, file.Name())
 	err = Validate(o)
 	assert.Error(t, err)
 	assert.Contains(t, err.Error(), "unable to load provider CA file(s)")